Actually enable response parameter encryption for unseal
This wasn't enabled because AttrResponseEncrypt was not passed to the session. Also stop setting a
random auth value in the sealed key object, as this breaks response parameter encryption during unseal
(as the encryption session is also the auth session, the sealed key object's auth value forms part of the
symmetric key).