keydata: use snapcore/snapd/osutil instead of renameio
This PR switches to use the snapd code to handle the writing of
the keydata. Adding extra dependencies (like renameio) in snapd
is problematic because it builds on all distros so we always
need to make sure the depdencies are available. For this use-case
it seems like our snapd code is equally well suited like renameio.