Improve the ProvisionTPM API
This makes some changes to the way that ProvisionTPM works, to enable it to be used on TPMs that have previously
been provisioned. This may be useful for recovery scenarios (where, eg, you might want to recreate the storage
root key without clearing the TPM).

It also returns more useful errors to assist the caller to respond appropriately to certain error conditions.