Revoke old authorization policies when new ones are created
We store the sealed key object and its associated metadata on disk. Updating the authorization policy currently
involves resealing the key object and then performing an atomic update of the file on disk. However, we don't
make any effort to prevent the old sealed key object with its old policy digest and associated metadata being
recovered from the disk even after the original file is unlinked, and there isn't really a 100% reliable way to
perform secure deletion of files anyway. Even if there was, there may be copies of the old sealed key object and
its associated metadata elsewhere.

One way to resolve this might be to persist the sealed key object and its associated metadata in the TPM. But
there are drawbacks with this approach:
- Updating the policy involves persisting a new object which requires owner authorization. We should avoid having
to take ownership of the TPM's storage hierarchy in order to perform normal operations, giving device owners the
option of taking ownership instead.
- Updating the policy requires a primary key and creating this also requires owner authorization. We currently
persist the primary key for this reason, but then this would mean we have to persist 2 objects.
- In addition to the primary key and sealed key object, we have other metadata associated with the key to support
the use of a PIN and to support the execution of OR policies. The complete size by default is close to 900bytes
already, which is a big chunk of the TPM's available NV storage (and this could be much larger for transient
policies that have multiple OR branches during an upgrade - it's entirely possible we could create transient
policies with metadata that doesn't fit in a single slot on the TPM).

NV counter indexes have an interesting set of properties that can help us here:
- They can only be incremented. There is no mechanism to decrement them.
- There is no mechanism to roll a counter back by undefining one and then redefining a new counter at the same
handle. New counters always start at a value higher than any previous counter value on the TPM.

Using this property of NV counters in combination with the TPM2_PolicyNV assertion allows us to implement policy
revocation by reserving a counter and adding an assertion in the authorization policy that the counter is less
than or equal to the current counter value + 1. Once the sealed key object and its associated metadata has been
atomically updated on disk, the counter is incremented which permanently revokes all authorization policies that
have been created previously.