Use the correct digest algorithm in each context
The defaultHashAlgorithm constant is used in some places, whilst tpm2.AlgorithmSHA256 is hardcoded. Although
everything is effectively hard coded to SHA-256 at the moment, introduce some more constants to make it
obvious which entities are related with respect to algorithm selection (eg, sealedKeyNameAlgorithm is used
for the policy session used for key unsealing, because the algorithm needs to match the name algorithm of the
sealed data object.