Implement PIN support using a NV index
The previous implementation used a second TPM object for PIN verification, with both parts of the object being
stored in the same file as the sealed key object and the rest of its metadata. Changing the PIN involved changing
the authorization value for this object, which recreates the private part of it. The key file was then updated
atomically on disk to incorporate the new private part.

As the PIN is changed without changing the name of the object, and therefore doesn't require any updates to the
authorization policy for the sealed key object, it is not possible to revoke older private parts of the TPM
object used for the PIN (which may include an old PIN or no PIN). We don't make any effort to prevent the old
PIN object with its old PIN value being recovered from the disk even after the original key file is unlinked, and
there isn't really a 100% reliable way to perform secure deletion of files anyway. Even if there was, there may be
copies of the old key file elsewhere.

This can be fixed by using a NV index on the TPM for PIN verification instead. Residing on the TPM, it is possible
to change an authorization value for a NV index without being able to rewind to an older version with a previous
authorization value. Lnowledge of the authorization value of the NV index is asserted using the PolicySecret
assertion (the same policy assertion used already), which includes the name of the NV index in the generated
authorization policy.

To protect against the TPM owner (the person who knows the storage hierachy authorization, which we leave empty by
default) from deleting the authorization value (ie, clearing the PIN) by undefining the NV index and then
redifining another one in its place with the same properties, we have the following requirements for the NV index
we create:
- It must be written to before its name is incorporated in to an authorization policy via PolicySecret.
- It must only be writable via a signed authorization policy.

This works because the name of the key used to sign the authorization for writing is included in the authorization
policy digest for the NV index. The authorization policy and attributes (including the one that indicates it has
been written to) are included in the name of the NV index. The name of the NV index is included in the
authorization policy digest for the sealed key object. Without the private part of the signing key used to
initialize the NV index, it is impossible for anybody to create another one with the same name.

As the NV index is zero sized and doesn't contain any other useful data, we throw away the signing key used to
perform the initial write for initializing it.