Add initial APIs for verifying a TPM's EK cert and verifying that we're connected to the TPM it was issued for
This adds SecureConnectToDefaultTPM which verifies the TPM's EK certificate against
a set of embedded manufacturer roots, and then verifies that we're connected to the
TPM for which the cert was issued by creating a session that's salted with a value
protected by the corresponding EK, and executing a command with that session.

Also added is SecureConnectToDefaultUnprovisionedTPM which does the same thing, but
doesn't verify that we're connected to the TPM for which the cert was issued, and
can therefore be used before the TPM is provisioned with a persistent EK. That
verification is deferred until ProvisionTPM is called.

The TPMConnection type is added, which provides access to the verified EK cert chain.
It also provides access to the ResourceContext associated with the verified EK, and
a HMAC session salted with a value protected by the EK associated with the verified
certificate.

ConnectToDefaultTPM also returns a TPMConnection with a EK ResourceContext (if the
TPM is provisisoned) and a session salted with a value protected by that EK. But in
this case, the EK isn't associated with a verifiable certificate and values protected
by the public part of it aren't guaranteed to remain confidential because we have to
read the key's public area and associated attributes from the TPM, and there's no
way to guarantee that the object on the TPM actually has the properties required of
an endorsement key (that it is restricted and non-duplicable).

This contains some unrelated changes because of an update to go-tpm2.