Re-add the ability to bind some data to the sealed key object
An earlier commit removed this functionality, but some auxiliary data is only
used for computing authorization policies and not used during policy execution,
meaning that it could be tampered with offline without causing an unlock
failure. Ensure this can be detected when computing new authorization policies
for an existing key data file.