Remove the owner auth fallback for reading the EK certificate
Section 3.6.1 of the "TCG PC Client Platform TPM Profile (PTP) Specification"
says that TPMA_NV_AUTHREAD shall be set and that the authorization value for
the EK cert index shall be the NULL value. Consider a non-null authorization
value or TPMA_NV_AUTHREAD being clear as broken, as requiring the owner
authorization during boot is problematic.