Integrity protect TPM2_Load in loadKeyData
When SealKeyToTPM is called to update an existing key, this guarantees that the
subsequent call to TPM2_Create seals the key to the same object.

Also make ProvisionTPM always refresh the SRK again, as it's not really possible
to detect whether an object on the TPM was created by us with our template.