Update the PCR allocation in the policy
We're not going to use the policy to lock down the entire GRUB configuration
now, as the main motivation for that was to lock down the kernel commandline
and there is a better way of doing that (via the systemd EFI stub for the
combined kernel / initramfs, which does the measurement before executing any
kernel code).

However, we don't want that to share a PCR with GRUB, else we would still have
to generate an authorization policy from the GRUB configuration. The kernel
commandline will instead be measured in to PCR12. To avoid taking up too many
PCRs for Ubuntu Core specific things, we'll also use the same one for measuring
properties of the snap model which is used to bind an encrypted drive to a
specific model.