Store the initialization key name for the pin NV index on disk, rather than storing policy data in the TPM
A previous change moved a policy digest in to the TPM which might be useful to share a single PIN NV index
for multiple sealed data objects, but it's not a requirement and so just store the name of the
initialization key on disk. This can be used to compute authorization policy OR digests in order to perform
PIN changes.