Make unlock call cryptsetup with a passphrase rather than master key
Traditional LUKS setups encrypt the volume key to one or more keyslots
using an intermediate key that's derived from a passphrase. In an ideal
world when sealing the disk encryption key to the TPM, the secret that
would be encrypted would be the intermediate key (so, the TPM would unseal
the intermediate key, and the intermediate key would be used to decrypt the
volume key from a keyslot).

However, keyslots in LUKS are really tied to passphrases. The KDFs that
derive the intermediate keys from passphrases are slow and resource
intensive, to harden the system against brute-forcing low-entropy
passphrases. As we can seal an arbitrarily long key to the TPM rather
than a low-entropy passphrase, we don't need the overhead of a resource
intensive KDF, and we don't want to block the boot for a couple of seconds
whilst the KDF derives the intermediate key from a secret unsealed from the
TPM. One option could have been to configure LUKS with a passphrase/keyslot
and then seal the intermediate key derived from the passphrase to the TPM,
but there is no support in cryptsetup to do this. For the initial plan for
Ubuntu Core, we opted to just seal the volume key to the TPM instead, which
can be used to open LUKS volumes in cryptsetup.

However, opening a volume with the volume key is not supported in
systemd-cryptsetup, which is what we want in the initramfs rather than the
full cryptsetup utility. To support systemd-cryptsetup, we can configure
the LUKS volume with a keyslot and high-entropy passphrase (say, a 64
byte key), but configure the KDF to execute with minimal overhead so that
it derives the intermediate key quickly.

This commit updates the unlock tool to open the LUKS volume with a
passphrase, which is really a high-entropy key.