Set the key size constraint to 32-bytes in SealKeyToTPM
The key size constraint was previously set to 64-bytes because we were
sealing the volume key, and 64-bytes is required by AES-XTS-256. Now we
aren't sealing the volume key, and the intermediate key derived from the
KDF and used to decrypt the volume key is 32-bytes. There is no point in
sealing secret that has more entropy than this, so adjust the constraint
accordingly.