Switch the authorization policy locking mechanism to measuring an event
This is closer to the original plan. The previous implementation worked by
read locking the PIN NV index, which changes its name until the next TPM
reset or restart and has the same effect. The problem with this is that
the NV index handle needs to be known by early-boot code. If someone boots
an unencrypted Ubuntu Core image from external media on a system that has
an encrypted writable partition, the early boot code running from the
external media won't be able to determine the handle of the PIN NV index.